Add support for Django 4.x
Change the deletion strategy from django.db.models.SET_NULL to
django.db.models.deletion.SET_NULL, in accordance with the
new package structure on Django 4.x.

This is a NOOP in terms of the DB schema as the strategies are
equivalent; it's just the path to that strategy class that has changed.